feat: extract svg components, utils function, and types
