+++ src/tags.web.ts
@@ -1,0 +1,60 @@
+import {
+  Circle,
+  ClipPath,
+  Defs,
+  Ellipse,
+  Filter,
+  FeColorMatrix,
+  FeGaussianBlur,
+  FeOffset,
+  ForeignObject,
+  G,
+  Line,
+  LinearGradient,
+  Marker,
+  Mask,
+  Path,
+  Pattern,
+  Polygon,
+  Polyline,
+  RadialGradient,
+  Rect,
+  Stop,
+  Svg,
+  Symbol,
+  Text,
+  TextPath,
+  TSpan,
+  Use,
+} from './web/elements';
+
+export const tags = {
+  circle: Circle,
+  clipPath: ClipPath,
+  defs: Defs,
+  ellipse: Ellipse,
+  filter: Filter,
+  feColorMatrix: FeColorMatrix,
+  feGaussianBlur: FeGaussianBlur,
+  feOffset: FeOffset,
+  foreignObject: ForeignObject,
+  g: G,
+  image: Image,
+  line: Line,
+  linearGradient: LinearGradient,
+  marker: Marker,
+  mask: Mask,
+  path: Path,
+  pattern: Pattern,
+  polygon: Polygon,
+  polyline: Polyline,
+  radialGradient: RadialGradient,
+  rect: Rect,
+  stop: Stop,
+  svg: Svg,
+  symbol: Symbol,
+  text: Text,
+  textPath: TextPath,
+  tspan: TSpan,
+  use: Use,
+} as const;
