+++ src/web/types.ts
@@ -1,0 +1,66 @@
+import type { ImageProps as RNImageProps } from 'react-native';
+import type {
+  NumberArray,
+  NumberProp,
+  TransformProps,
+} from '../lib/extract/types';
+
+type BlurEvent = object;
+type FocusEvent = object;
+type PressEvent = object;
+type LayoutEvent = object;
+type EdgeInsetsProp = object;
+
+export interface BaseProps {
+  accessible?: boolean;
+  accessibilityLabel?: string;
+  accessibilityHint?: string;
+  accessibilityIgnoresInvertColors?: boolean;
+  accessibilityRole?: string;
+  accessibilityState?: object;
+  delayLongPress?: number;
+  delayPressIn?: number;
+  delayPressOut?: number;
+  disabled?: boolean;
+  hitSlop?: EdgeInsetsProp;
+  href?: RNImageProps['source'] | string | number;
+  nativeID?: string;
+  touchSoundDisabled?: boolean;
+  onBlur?: (e: BlurEvent) => void;
+  onFocus?: (e: FocusEvent) => void;
+  onLayout?: (event: LayoutEvent) => object;
+  onLongPress?: (event: PressEvent) => object;
+  onClick?: (event: PressEvent) => object;
+  onPress?: (event: PressEvent) => object;
+  onPressIn?: (event: PressEvent) => object;
+  onPressOut?: (event: PressEvent) => object;
+  pressRetentionOffset?: EdgeInsetsProp;
+  rejectResponderTermination?: boolean;
+
+  transform?: TransformProps['transform'];
+  translate?: NumberArray;
+  translateX?: NumberProp;
+  translateY?: NumberProp;
+  scale?: NumberArray;
+  scaleX?: NumberProp;
+  scaleY?: NumberProp;
+  rotation?: NumberProp;
+  skewX?: NumberProp;
+  skewY?: NumberProp;
+  origin?: NumberArray;
+  originX?: NumberProp;
+  originY?: NumberProp;
+
+  fontStyle?: string;
+  fontWeight?: NumberProp;
+  fontSize?: NumberProp;
+  fontFamily?: string;
+  forwardedRef?:
+    | React.RefCallback<SVGElement>
+    | React.MutableRefObject<SVGElement | null>;
+  style?: Iterable<unknown>;
+
+  // different tranform props
+  gradientTransform?: TransformProps['transform'];
+  patternTransform?: TransformProps['transform'];
+}
